task/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-ckeditor==5.9.0
django-cors-headers==3.8.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
idna==2.8
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0